Silence C4244 in CCedtDoc::GetFileSize()
CFileStatus::m_size widened to ULONGLONG in modern MFC, so the
implicit narrowing to LONG triggers C4244. Make the truncation
explicit with static_cast<LONG>; the 4 GB ceiling matches the
existing call sites which already accept a DWORD.

+++ src/doc/cedtDoc.h
@@ -102,7 +102,7 @@ public: // *** cedtDoc.cpp ***
 	CString GetRemotePathName() const { return m_szRemotePathName; }
 
 	CString GetFullRemotePathName() const;
-	LONG GetFileSize() const { return m_clsFileStatus.m_size; }
+	LONG GetFileSize() const { return static_cast<LONG>(m_clsFileStatus.m_size); }
 
 protected:
 //	BOOL OnNewDocument();
